// src/app/enums/category.ts
/**
 * Enumeration with the types of categories
 *
 * @author Jhonier Gaviria M.
 * @version 0.1.0
 */
export enum Category {
  ARITHMETIC_OPERATOR = 'Operador Arímetico',
  ASSIGNMENT_OPERATOR = 'Operador Asignación',
  BLOCK_COMMENT = 'Comentario de bloque',
  CHARACTER = 'Carácter',
  DECIMAL = 'Decimal',
  DECREMENT_OPERATOR = 'Operador Decremento',
  DIFFERENT_OPERATOR = 'Operador Diferente',
  DIVISION_OPERATOR = 'Operador división',
  DIVISION_ASSIGNMENT_OPERATOR = 'Operador asignación de división',
  END_OF_SENTENCE_OPERATOR = 'Operador fin de sentencia',
  EQUALS_OPERATOR = 'Operador Igual',
  IDENTIFIER = 'Identificador',
  INCREMENT_OPERATOR = 'Operador Incremento',
  INTEGER = 'Entero',
  LINE_COMMENT = 'Comentario de linea',
  LOGIC_OPERATOR = 'Operador Lógico',
  MINOR_OPERATOR = 'Operador Menor',
  MINOR_OR_EQUAL_OPERATOR = 'Operador Menor o Igual',
  MAJOR_OPERATOR = 'Operador Mayor',
  MAJOR_OR_EQUAL_OPERATOR = 'Operador Mayor o Igual',
  MODULO_OPERATOR = 'Operador módulo',
  MODULO_ASSIGNMENT_OPERATOR = 'Operador asignación de módulo',
  MULTIPLICATION_OPERATOR = 'Operador multiplicación',
  STRING = 'Cadena de caracteres',
  MULTIPLICATION_ASSIGNMENT_OPERATOR = 'Operador asignación de multiplicación',
  SUBTRACTION_OPERATOR = 'Operador resta',
  SUBTRACTION_ASSIGNMENT_OPERATOR = 'Operador asignación de resta',
  SUM_OPERATOR = 'Operador suma',
  SUM_ASSIGNMENT_OPERATOR = 'Operador asignación de suma',
  UNKNOWN = 'Desconocido'
}